refactor: move header related components to the same level
