+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
 
-import Header from '@/components/Header/Header';
+import Header from '@/components/header/Header';
 import Footer from '@/components/footer/Footer';
 import { AuthProvider } from '@/context/AuthContext';
 const geistSans = Geist({
