+++ src/components/header/ProfileCard.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Image from 'next/image';
 import { mockedAvatar } from '@/assets';
 import { useAuth } from '@/context/AuthContext';
-import ProfileMenu from '@/components/Header/ProfileCard/ProfileMenu';
+import ProfileMenu from '@/components/header/ProfileMenu';
 import { useState } from 'react';
 
 const ProfileCard = () => {
